modularized cross-validation function code
